Allow user without config and authentication.
Create a seeder command which will automatically store the default configurations in database.

Fix the error which was generated while config key does not exist on the database

Store the username and password in the file instead of database.

Set middleware if required config not set then ask to set it forcefully.